requirements.txt
py-cord
aiohttp
joblib
scikit-learn
numpy
pandas
python-dotenv